requirements.txt
chardet
lunarcalendar
deep-translator
speedtest-cli
psutil
google-generativeai
pygments
moviepy
selenium
gTTS
langdetect
pytesseract
wikipedia
matplotlib
forex-python
tqdm
py-cpuinfo
googletrans==4.0.0-rc1
openai-whisper
absl-py==2.2.2
aenum==2.2.6
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
anytree==2.13.0
astunparse==1.6.3
attr==0.3.2
attrs==25.1.0
audioread==3.0.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
decorator==4.4.2
deep-translator==1.11.4
dmidecode==0.9.0
duckduckgo_search==7.3.2
easyocr==1.7.2
emoji==2.14.1
ephem==4.2
fake-useragent==2.0.3
fbchat==1.9.7
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
forex-python==1.8
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gast==0.6.0
google==3.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.9.0
google-pasta==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
graphviz==0.20.3
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
h5py==3.13.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
jellyfish==1.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.10.0
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
LunarCalendar==0.0.9
lxml==5.3.0
lyricsgenius==3.2.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munch==4.0.0
namex==0.0.9
networkx==3.4.2
ninja==1.11.1.4
nltk==3.8.1
nsfw-detector==1.1.1
nudenet==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.22.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
paho-mqtt==1.6.1
pillow==10.4.0
pip-autoremove==0.10.0
pipdeptree==2.26.1
platformdirs==4.3.6
pooch==1.8.2
primp==0.12.0
proglog==0.1.10
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydub==0.25.1
pyfiglet==1.0.2
Pygments==2.19.1
pyjokes==0.8.3
PyMatting==1.1.13
pyparsing==3.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.1
referencing==0.36.2
regex==2024.11.6
rel==0.4.9.19
rembg==2.0.65
removebg==0.4
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
setuptools==59.5.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.14.3
speedtest-cli==2.1.3
sympy==1.13.1
Telethon==1.38.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.5.0
tifffile==2025.3.30
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
unicode==2.9
Unidecode==1.3.8
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
youtube-search==2.1.2
youtube-search-python
yt-dlp==2025.1.26
zlapi==1.0.3
zope.interface==7.2
